Finally figured out how to properly use my different pages correctly.
