+++ Project/tinyTask/HomeScreen.js
@@ -1,10 +1,25 @@
 import React from "react";
-import { View, Text } from "react-native";
+import { StyleSheet, SafeAreaView, Text } from "react-native";
 
 export default function HomeScreen() {
   return (
-    <View>
-      <Text>Home Screen</Text>
-    </View>
+    <SafeAreaView style={styles.container}>
+      <Text style={styles.h1}>Home Screen</Text>
+    </SafeAreaView>
   );
 }
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: "center",
+    backgroundColor: "#ecf0f1",
+    padding: 8,
+  },
+  h1: {
+    margin: 28,
+    fontSize: 36,
+    fontWeight: "bold",
+    textAlign: "center",
+  },
+});
